+++ apps/web/components/Features/LeaveFeature.tsx
@@ -28,7 +28,7 @@ type Post struct {
 `;
 
 const bye = `
-$ rm -rf .zapp
+$ rm -rf .specui
 `;
 
 export const LeaveFeature = () => {
@@ -36,8 +36,8 @@ export const LeaveFeature = () => {
     <div className="pb-32 px-4 mx-auto max-w-6xl">
       <FeatureTitle>Leave anytime</FeatureTitle>
       <FeatureDescription>
-        If you ever want to stop using SpecUI, just remove the `.zapp` directory and keep shipping.
-        We will miss you and are honored to be a part of your journey.
+        If you ever want to stop using SpecUI, just remove the `.specui` directory and keep
+        shipping. We will miss you and are honored to be a part of your journey.
       </FeatureDescription>
       <div className="gap-4 grid grid-cols-3">
         <CodeSnippet className="col-span-3 md:col-span-1" title="spec.yml">