+++ generators/vanilla/.specui/generator-browser.ts
@@ -1,0 +1,7 @@
+import { PrettierProcessor } from '@specui/prettier/dist/standalone';
+import { ISpec } from './interfaces/ISpec';
+import generator from './generator';
+
+export default async function generatorBrowser(spec: ISpec) {
+  return generator(spec, PrettierProcessor);
+}