+++ generators/vanilla/.specui/generator-node.ts
@@ -1,0 +1,7 @@
+import { PrettierProcessor } from '@specui/prettier';
+import { ISpec } from './interfaces/ISpec';
+import generator from './generator';
+
+export default async function generatorNode(spec: ISpec) {
+  return generator(spec, PrettierProcessor);
+}