+++ examples/next-api/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Patch Changes
 
-- @specui/next-zapp@0.0.1
+- @specui/next-generator@0.0.1